requirements.txt
# Dependencias core (obligatorias)
requests>=2.28.0
lxml>=4.9.0
pyarrow>=8.0.0
pandas>=1.5.0
python-dotenv>=0.21.0

# Dependencias para web scraping
beautifulsoup4>=4.13.4

# Dependencias para S3
boto3>=1.26.0

# Dependencias para Pinecone
pinecone>=6.0.0

# Otras dependencias
python-xbrl
psutil
tqdm

# MongoDB y Motor - Versiones compatibles
pymongo==4.5.0
motor==3.3.1

# Nota: zipfile36 fue reemplazado por zipfile en la biblioteca estándar de Python

# New dependencies
pydantic>=2.0.0
pydantic-settings>=2.0.0
tiktoken>=0.3.0
openai>=1.0.0

# Stack Async
fastapi==0.109.2
celery==5.3.6
redis==5.0.1
aiohttp==3.9.3
dask==2024.1.1
pinecone-client==3.0.2
prometheus-client==0.19.0
uvicorn==0.27.1
python-multipart==0.0.9
aiofiles==23.2.1
httpx==0.26.0